+++ src/components/Form/FormWrapper/FormWrapper.styled.tsx
@@ -3,6 +3,11 @@ import styled from "styled-components";
 import { FormItemStyled } from "../FormItem/FormItem.styled";
 import { FormItemGroupStyled } from "../FormItemGroup/FormItemGroup.styled";
 
+import { CheckBoxStyled } from "../CheckBox/CheckBox.styled";
+import { InputFieldStyled } from "../InputField/InputField.styled";
+import { SelectStyled } from "../Select/Select.styled";
+import { TextAreaStyled } from "../TextArea/TextArea.styled";
+import { FormWrapperProps } from "./FormWrapper";
 
 export const FormWrapperStyled = styled.form`
     display        : flex;
@@ -18,7 +23,7 @@ export const FormWrapperStyled = styled.form`
     &.spacing-medium ${FormItemStyled}:not(:last-child) { margin-bottom : 32px; }
     &.spacing-large  ${FormItemStyled}:not(:last-child) { margin-bottom : 40px; }
     &.spacing-huge   ${FormItemStyled}:not(:last-child) { margin-bottom : 48px; }
-
+/* 
     &.spacing-none   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom :    0; }
     &.spacing-nano   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom :  8px; }
     &.spacing-micro  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 12px; }
@@ -26,5 +31,5 @@ export const FormWrapperStyled = styled.form`
     &.spacing-small  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 24px; }
     &.spacing-medium ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 32px; }
     &.spacing-large  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 40px; }
-    &.spacing-huge   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 48px; }
+    &.spacing-huge   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 48px; } */
 `;