+++ src/components/Form/Select/Select.tsx
@@ -4,9 +4,7 @@ import { Element } from "../../Element/Element";
 import { CommonAndHTMLProps } from "../../Element/constants";
 
 import { SelectWrapperStyled, SelectStyled } from "./Select.styled";
-
-import { InputLabel } from "../InputLabel/InputLabel";
-import { FormItem } from "../FormItem/FormItem";
+import { BaseInputComponent } from "../BaseInputComponent/BaseInputComponent";
 
 export type SelectElementType = HTMLSelectElement;
 export type OptionElementType = HTMLOptionElement;
@@ -22,28 +20,20 @@ export interface SelectCustomProps {
 export type OptionProps = CommonAndHTMLProps<OptionElementType>;
 export type SelectProps = Omit<CommonAndHTMLProps<SelectElementType>, keyof SelectCustomProps> & SelectCustomProps;
 
-const Option = ({ name, ...props }: OptionProps) => {
+const SelectWithOptions = ({ options, className, ...props }: SelectProps) => {
     return (
-        <Element<OptionElementType> as="option" {...props}>
-            {name}
+        <Element<HTMLDivElement> as={SelectWrapperStyled} className={className}>
+            <Element<SelectElementType> as={SelectStyled} {...props}>
+                {options.map((option, index) => (
+                    <Element<OptionElementType> as="option" key={index} {...option}>
+                        {option.name}
+                    </Element>
+                ))}
+            </Element>
         </Element>
     );
 };
 
-export const Select = React.forwardRef(
-    ({ label, helpText, errorText, className, options, ...props }: SelectProps, ref: React.Ref<SelectElementType>) => {
-        return (
-            <FormItem>
-                <Element<HTMLDivElement> as={SelectWrapperStyled} className={className}>
-                    <Element<SelectElementType> as={SelectStyled} ref={ref} {...props}>
-                        {options.map((option, index) => {
-                            return <Option key={index} {...option} />;
-                        })}
-                    </Element>
-                </Element>
-
-                {label && <InputLabel label={label} helpText={helpText} errorText={errorText} />}
-            </FormItem>
-        );
-    }
-);
+export const Select = React.forwardRef((props: SelectProps, ref: React.Ref<SelectElementType>) => {
+    return <BaseInputComponent<SelectElementType> as={SelectWithOptions} ref={ref} {...props} />;
+});